feat(cmd): use cobra for CLI parsing
