--- cmd/notary/main.go
@@ -1,81 +1,0 @@
-package main
-
-import (
-	"flag"
-	"log"
-	"net/http"
-	"os"
-	"os/signal"
-
-	"github.com/canonical/notary/internal/config"
-	"github.com/canonical/notary/internal/db"
-	"github.com/canonical/notary/internal/server"
-	"go.uber.org/zap"
-)
-
-func main() {
-	appContext, err := parseCLIInput()
-	l := appContext.Logger
-	if err != nil {
-		log.Fatalf("couldn't create app context: %s", err)
-	}
-
-	// Initialize the database connection
-	db, err := db.NewDatabase(&db.DatabaseOpts{
-		DatabasePath: appContext.DBPath,
-		Backend:      appContext.EncryptionBackend,
-		Logger:       appContext.Logger,
-	})
-	if err != nil {
-		l.Fatal("couldn't initialize database", zap.Error(err))
-	}
-
-	// Initialize and run the API and webserver
-	srv, err := server.New(&server.ServerOpts{
-		TLSCertificate:            appContext.TLSCertificate,
-		TLSPrivateKey:             appContext.TLSPrivateKey,
-		Port:                      appContext.Port,
-		Database:                  db,
-		ExternalHostname:          appContext.ExternalHostname,
-		EnablePebbleNotifications: appContext.PebbleNotificationsEnabled,
-		Logger:                    appContext.Logger,
-		PublicConfig:              appContext.PublicConfig,
-	})
-	if err != nil {
-		l.Fatal("couldn't initialize server", zap.Error(err))
-	}
-	l.Info("Starting server at", zap.String("url", srv.Addr))
-	if err := srv.ListenAndServeTLS("", ""); err != http.ErrServerClosed {
-		l.Fatal("HTTP server ListenAndServe", zap.Error(err))
-	}
-	l.Info("Shutting down server")
-
-	// Listen for SIGINT to begin graceful shutdown
-	mainThread := make(chan struct{})
-	go func() {
-		sigint := make(chan os.Signal, 1)
-		signal.Notify(sigint, os.Interrupt)
-		<-sigint
-		l.Info("interrupt signal received")
-		close(mainThread)
-	}()
-
-	// Await sigint listener to release main thread
-	<-mainThread
-	l.Info("server shutdown completed.")
-}
-
-// Parses the command line arguments and returns the appropriate app context.
-func parseCLIInput() (*config.NotaryAppContext, error) {
-	log.SetOutput(os.Stderr)
-	configFilePtr := flag.String("config", "", "The config file to be provided to the server")
-	flag.Parse()
-	if *configFilePtr == "" {
-		log.Fatalf("providing a config file is required.")
-	}
-	app, err := config.CreateAppContext(*configFilePtr)
-	if err != nil {
-		log.Fatalf("couldn't create app context: %s", err)
-	}
-	return app, nil
-}
