+++ docs/reference/config_file.md
@@ -2,7 +2,7 @@
 
 Notary is configured using a YAML file.
 
-Start Notary with the `--config` flag to specify the path to the configuration file.
+Start Notary with the `start` command and `--config` flag to specify the path to the configuration file.
 Or If you are using the snap you can modify the config under `/var/snap/notary/common/notary.yaml`
 
 ## Parameters
