+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/canonical/notary/cmd"
+
+func main() {
+	cmd.Execute()
+}
